backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
boto3>=1.34.120
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
